+++ pages/_app.tsx
@@ -1,8 +1,14 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 
+import Auth from '../components/auth'
+
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return (
+    <Auth>
+      <Component {...pageProps} />
+    </Auth>
+  )
 }
 
 export default MyApp
